use shareable fixtures for e2e tests and setup e2e example
